requirements.txt
django==3.0.8
gunicorn==20.0.4
psycopg2-binary==2.8.5
requests==2.24.0
celery==4.4.6
ipdb==0.13.3
djangorestframework==3.11.0
django-admin-rangefilter==0.6.1
coverage==5.2
python-decouple==3.3